+++ Archive Parsers/Unix/Archivers/XADTarSparseHandle.m
@@ -234,7 +234,7 @@
 // 		fprintf( stderr, "Reading: %d really %d.\n", positionInBuffer, currentOffset );
 		if( regions[ currentOffset ].hasData )
 		{
-			[parent readAtMost:dataLeftInRegion toBuffer:buffer];
+			[parent readAtMost:(int)dataLeftInRegion toBuffer:buffer];
 		}
 		currentRegion = regions[ currentRegion ].nextRegion;
 		positionInRegion = 0;
@@ -245,7 +245,7 @@
 	// Read the last segment of data, if required.
 	if( regions[ currentRegion ].hasData )
 	{
-		[parent readAtMost:(num - positionInBuffer) toBuffer:buffer];
+		[parent readAtMost:(int)(num - positionInBuffer) toBuffer:buffer];
 		currentOffset += num - positionInBuffer;
 	}
 
@@ -266,7 +266,7 @@
 	
 // 	fprintf( stderr, "Readatmost okay, read %d.\n", positionInBuffer );
 
-	return( positionInBuffer );
+	return( (int)positionInBuffer );
 }
 
 @end
